# api-gateway/api/docs/swagger.yaml
definitions:
  booking.BookingCreateReq:
    properties:
      location:
        type: string
      provider_id:
        type: string
      scheduled_time:
        type: string
      service_id:
        type: string
      status:
        type: string
      total_price:
        type: integer
      user_id:
        type: string
    type: object
  booking.BookingGetAllRes:
    properties:
      bookings:
        items:
          $ref: '#/definitions/booking.BookingRes'
        type: array
    type: object
  booking.BookingGetByIdRes:
    properties:
      booking:
        $ref: '#/definitions/booking.BookingRes'
    type: object
  booking.BookingRes:
    properties:
      id:
        type: string
      location:
        type: string
      provider_id:
        type: string
      scheduled_time:
        type: string
      service_id:
        type: string
      status:
        type: string
      total_price:
        type: integer
      user_id:
        type: string
    type: object
  booking.NotificationGetAllRes:
    properties:
      notification:
        items:
          $ref: '#/definitions/booking.NotificationRes'
        type: array
    type: object
  booking.NotificationGetByIdRes:
    properties:
      notification:
        $ref: '#/definitions/booking.NotificationRes'
    type: object
  booking.NotificationRes:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_read:
        type: string
      message:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /booking:
    post:
      consumes:
      - application/json
      description: Create a new booking
      parameters:
      - description: Booking data
        in: body
        name: booking
        required: true
        schema:
          $ref: '#/definitions/booking.BookingCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: Booking created
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create booking
      tags:
      - booking
  /booking/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a booking by ID
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Booking deleted
          schema:
            type: string
        "400":
          description: Invalid booking ID
          schema:
            type: string
        "404":
          description: Booking not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete booking
      tags:
      - booking
    get:
      consumes:
      - application/json
      description: Get a booking
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get booking
      tags:
      - booking
    put:
      consumes:
      - application/json
      description: Update an existing booking
      parameters:
      - description: Id
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Booking updated
          schema:
            type: string
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Booking not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update booking
      tags:
      - booking
  /booking/all:
    get:
      consumes:
      - application/json
      description: Get all booking
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.BookingGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all booking
      tags:
      - booking
  /notification/{id}/read:
    get:
      consumes:
      - application/json
      description: Get a notification
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.NotificationGetByIdRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get notification
      tags:
      - notification
  /notification/all:
    get:
      consumes:
      - application/json
      description: Get all notification
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: IsRead
        in: query
        name: is_read
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/booking.NotificationGetAllRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all notification
      tags:
      - notification
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"